feat: Add DownloadManager Support
Currently using hardcoded string, makes a request to DownloadManager to download the file at the specified URL.  DownloadProgress.kt (should rename and add onReceive log to the class) accurately displays the download progress and allows us to display it to the screen.  This will help us when we add the custom button to change the animation based on download progress.