feat: Added notification support
DownloadManager file is shown while downloading to show the user the download progress when outside of the app.  It is set to disappear when the download is done and then we display our own notification, which will take the user to the DetailFragment when clicked.

refactor: DownloadProgress.kt renamed to DownloadUtils.kt and reworked to use Flow.

Switching to flow allowed us to move the isDownloading logic out of the ViewModel and into the DownloadUtils.kt file.  Since we're using Flow now, we can loop through checking the progress until the download is finished and emit the value every pass of the loop.