feat: Basic Implementation for ButtonState
Created sealed class ButtonState to track the state of the button (from Disabled to Completed/Failed) and tested that updating state draws the intended colors/shapes to the canvas based on ButtonState. I decided to pause implementing this feature to look over the current animation code to see if it can be refactored/simplified to make it easier to animate both state and progress simultaneously.