feat: ButtonState expanded
Working code in place to update the view to the correct button state and animate between all states. (Currently ButtonState.Downloading doesn't show an icon, but that will be fixed soon.)  Adding a commit before completely rewritting the animation code and moving it into the Custom View instead of in DownloadFragment.kt.  (If I were to reuse the view in other parts, it'd be more logical for the view to be able to animate itself.  It took some trial and error to get through the animation code, and now that I have a better grasp on it, I can move it into the view, when earlier I didn't know how to implement that.)